separate ColorPalette from main theme file
